Converted size_t => off_t for all things involving file lengths and offsets

+++ src/chunking.cxx
@@ -11,9 +11,9 @@ namespace pcat
 	struct mappingOffset_t final
 	{
 	private:
-		size_t offset_{0};
-		size_t adjust_{0};
-		size_t length_{0};
+		off_t offset_{0};
+		off_t adjust_{0};
+		off_t length_{0};
 
 		constexpr void recalcAdjustment()
 		{
@@ -23,18 +23,18 @@ namespace pcat
 
 	public:
 		constexpr mappingOffset_t() noexcept = default;
-		constexpr mappingOffset_t(const size_t offset) noexcept : offset_{offset} { recalcAdjustment(); }
-		constexpr mappingOffset_t(const size_t offset, const size_t length) noexcept :
+		constexpr mappingOffset_t(const off_t offset) noexcept : offset_{offset} { recalcAdjustment(); }
+		constexpr mappingOffset_t(const off_t offset, const off_t length) noexcept :
 			offset_{offset}, length_{length} { recalcAdjustment(); }
-		[[nodiscard]] constexpr size_t offset() const noexcept { return offset_ + adjust_; }
-		[[nodiscard]] constexpr size_t adjustment() const noexcept { return adjust_; }
-		[[nodiscard]] constexpr size_t length() const noexcept { return length_; }
-		[[nodiscard]] constexpr size_t adjustedOffset() const noexcept { return offset_; }
-		[[nodiscard]] constexpr size_t adjustedLength() const noexcept { return length_ + adjust_; }
-		constexpr void length(const size_t length) noexcept { length_ = length; }
-		constexpr operator size_t() const noexcept { return offset(); }
+		[[nodiscard]] constexpr off_t offset() const noexcept { return offset_ + adjust_; }
+		[[nodiscard]] constexpr off_t adjustment() const noexcept { return adjust_; }
+		[[nodiscard]] constexpr off_t length() const noexcept { return length_; }
+		[[nodiscard]] constexpr off_t adjustedOffset() const noexcept { return offset_; }
+		[[nodiscard]] constexpr off_t adjustedLength() const noexcept { return length_ + adjust_; }
+		constexpr void length(const off_t length) noexcept { length_ = length; }
+		constexpr operator off_t() const noexcept { return offset(); }
 
-		constexpr mappingOffset_t &operator +=(const size_t amount) noexcept
+		constexpr mappingOffset_t &operator +=(const off_t amount) noexcept
 		{
 			offset_ += amount + adjust_;
 			recalcAdjustment();
@@ -42,16 +42,16 @@ namespace pcat
 		}
 	};
 
-	constexpr size_t blockLength(const size_t length)
+	constexpr off_t blockLength(const off_t length)
 		{ return std::min(transferBlockSize, length); }
 
 	void calculateInputChunking() noexcept
 	{
 		auto file = inputFiles.begin();
-		size_t inputLength = file->length();
+		off_t inputLength = file->length();
 		mappingOffset_t inputOffset{0, blockLength(inputLength)};
 
-		const size_t outputLength = outputFile.length();
+		const off_t outputLength = outputFile.length();
 		for (mappingOffset_t outputOffset{}; outputOffset < outputLength; outputOffset += transferBlockSize)
 		{
 			outputOffset.length(blockLength(outputLength - outputOffset));
@@ -72,10 +72,10 @@ namespace pcat
 
 			if (outputOffset.length() != inputOffset.length())
 			{
-				const size_t originalOutputOffset = outputOffset.offset();
+				const off_t originalOutputOffset = outputOffset.offset();
 				while (outputOffset.length() - inputOffset.length())
 				{
-					const size_t remainder = outputOffset.length() - inputOffset.length();
+					const off_t remainder = outputOffset.length() - inputOffset.length();
 					console.info("Transfer caused a remainder of ", remainder, " bytes to go for output block");
 					outputOffset += inputOffset.length();
 					outputOffset.length(remainder);
